# coprocessor: handle paging result even if not enabled